requirements.txt
pandas>=2.0.0
numpy>=1.26.0
requests>=2.32.0
python-dotenv>=1.0.0
